+++ src/hooks/useFieldColor.ts
@@ -1,0 +1,15 @@
+import { FieldPath, FieldValues, useController } from 'react-hook-form';
+
+export const useFieldColor = <Values extends FieldValues>(
+  name: FieldPath<Values>,
+  showDirty?: boolean,
+) => {
+  const {
+    fieldState: { error, isDirty },
+  } = useController({ name });
+  return error === undefined
+    ? isDirty && showDirty
+      ? 'success'
+      : 'ghost'
+    : 'error';
+};